backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.19
anthropic==0.42.0
pymupdf==1.25.1
python-docx==1.1.2
httpx==0.28.1
